Adding last file and updates of tghe readme file

+++ data_augmentation_TS.py
@@ -10,8 +10,6 @@ from tqdm import tqdm
 
 from utils import log_mels
 
-# tutorial considered
-
 def get_spectrogram(
         waveform,
         n_fft=400,
@@ -93,7 +91,7 @@ if __name__== "__main__":
     #os.makedirs(test_folder, exist_ok=True)
     
     # parameers
-    mel_bands = 64
+    mel_bands = 128
     target_sample_rate = 16000
     num_samples = target_sample_rate * 4 # sample rate * audio_max_lenght
     n_window = 1204
@@ -110,6 +108,7 @@ if __name__== "__main__":
     #paths_list = paths_list[:10]
     
     time_stretch_values = [0.81, 0.93, 1.07, 1.23] 
+    ts = 0.81
     
     mel_spectogram = MelSpectrogram(
             sample_rate=target_sample_rate, 
@@ -130,41 +129,47 @@ if __name__== "__main__":
         # Time stretch is applied to the complex-valued spectgram 
         ##
         
-        # selecting a random values for the time stretch
-        time_stretch_selected = random.choice(time_stretch_values)
+        
+        #time_stretch_selected = random.choice(time_stretch_values)
         #print(f"Random values selected for the time stretch: {time_stretch_selected}")
         
         # spectogram
         signal_spectogram = get_spectrogram(signal, power=None)
         
-        # time stretch 
-        time_stretch = TimeStretch()
-        signal_time_strectched = time_stretch(signal_spectogram, overriding_rate=time_stretch_selected)
-        
-        # get back to audiowave
-        waveform_strecthed  = InverseSpectrogram()(signal_time_strectched)
-        
-        # save it to listen to it
-        #file_TS_signal = os.path.join(test_folder, f'file_{index}_TS_{time_stretch_selected}_{mel_bands}.wav')
-        #torchaudio.save(file_TS_signal, waveform_strecthed, target_sample_rate)
-        
-        # pad to be sure we got for seconds of the signal
-        waveform_strecthed = pad_to(waveform_strecthed, num_samples)
+        for time_stretch_selected in time_stretch_values[0:1]:
 
-        # log-mel spectogram
-        signal_TS_mel = mel_spectogram(waveform_strecthed)
-        signal_TS_mel = log_mels(signal_TS_mel, 'GPU')
-        
-        # save the plot time stretched
-        #plot_file_path_TS = os.path.join(test_folder, f'file_{index}_TS_{time_stretch_selected}_{mel_bands}.png')
-        #plot_figure(signal_TS_mel[0].detach().numpy(), plot_file_path_TS)
-        
-        # the file will need to be saved as numpy
-        file_audio_TS_path = paths_list[index].replace('.wav', f'_TS_{time_stretch_selected}_{mel_bands}.npy')
-        index_file = paths_list[index].split('/')[-1]        
-        annotations_augmented.loc[annotations_augmented['slice_file_name'] == index_file, 'slice_file_name'] = file_audio_TS_path
-        np.save(file_audio_TS_path, signal_TS_mel.detach().numpy())
+            file_audio_TS_path = paths_list[index].replace('.wav', f'_TS_{time_stretch_selected}_{mel_bands}.npy')
+            
+            if not os.path.exists(file_audio_TS_path):
+                # time stretch 
+                time_stretch = TimeStretch()
+                signal_time_strectched = time_stretch(signal_spectogram, overriding_rate=time_stretch_selected)
+                
+                # get back to audiowave
+                waveform_strecthed  = InverseSpectrogram()(signal_time_strectched)
+                
+                # save it to listen to it
+                #file_TS_signal = os.path.join(test_folder, f'file_{index}_TS_{time_stretch_selected}_{mel_bands}.wav')
+                #torchaudio.save(file_TS_signal, waveform_strecthed, target_sample_rate)
+                
+                # pad to be sure we got for seconds of the signal
+                waveform_strecthed = pad_to(waveform_strecthed, num_samples)
 
+                # log-mel spectogram
+                signal_TS_mel = mel_spectogram(waveform_strecthed)
+                signal_TS_mel = log_mels(signal_TS_mel, 'GPU')
+                
+                # save the plot time stretched
+                #plot_file_path_TS = os.path.join(test_folder, f'file_{index}_TS_{time_stretch_selected}_{mel_bands}.png')
+                #plot_figure(signal_TS_mel[0].detach().numpy(), plot_file_path_TS)
+                
+                # the file will need to be saved as numpy
+                #file_audio_TS_path = paths_list[index].replace('.wav', f'_TS_{time_stretch_selected}_{mel_bands}.npy')
+                np.save(file_audio_TS_path, signal_TS_mel.detach().numpy())
+            
+            index_file = paths_list[index].split('/')[-1]        
+            annotations_augmented.loc[annotations_augmented['slice_file_name'] == index_file, 'slice_file_name'] = file_audio_TS_path
+            
        
         # original audio 
         #original_signal_path = os.path.join(test_folder, f'file_{index}_{mel_bands}.wav')
@@ -178,7 +183,7 @@ if __name__== "__main__":
         # plot_figure(original_signal[0], plot_file_path)
         
     # save the metadata
-    augmented_TS_annotations = f"/nas/home/fronchini/EUSIPCO/urban-sound-class/UrbanSound8K/metadata/UrbanSound8K_TS_{mel_bands}.csv"
+    augmented_TS_annotations = f"/nas/home/fronchini/EUSIPCO/urban-sound-class/UrbanSound8K/metadata/UrbanSound8K_TS_{mel_bands}_{ts}.csv"
     annotations_augmented.to_csv(augmented_TS_annotations, index=False)
